+++ code/sorting/src/cycle_sort/cycle_sort.cpp
@@ -8,15 +8,15 @@ void cycleSort (vector<int>& arr)
 {
     // traverse array elements and put it to on
     // the right place
-    for (int cycle_start=0; cycle_start<=arr.size()-2; cycle_start++)
+    for (size_t cycle_start=0; cycle_start<=arr.size()-2; cycle_start++)
     {
         // initialize item as starting point
         int item = arr[cycle_start];
 
         // Find position where we put the item. We basically
         // count all smaller elements on right side of item.
-        int pos = cycle_start;
-        for (int i = cycle_start+1; i<arr.size(); i++)
+        size_t pos = cycle_start;
+        for (size_t i = cycle_start+1; i<arr.size(); i++)
             if (arr[i] < item)
                 pos++;
 
@@ -40,7 +40,7 @@ void cycleSort (vector<int>& arr)
             pos = cycle_start;
 
             // Find position where we put the element
-            for (int i = cycle_start+1; i<arr.size(); i++)
+            for (size_t i = cycle_start+1; i<arr.size(); i++)
                 if (arr[i] < item)
                     pos += 1;
 
@@ -70,7 +70,7 @@ int main()
     cycleSort(arr);
 
     cout << "\nAfter sort : " <<endl;
-    for(int i =0; i<arr.size(); i++)
+    for(size_t i =0; i<arr.size(); i++)
         cout << arr[i] << " ";
     return 0;
 }
